pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "firebase-admin>=7.1.0",
    "google-cloud-firestore>=2.21.0",
    "jinja2>=3.1.6",
    "openai>=1.101.0",
    "pydantic>=2.11.7",
    "pypdf2>=3.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.35.0",
]
